Pie chart, Custom Button, TotalRevenue and creating the From
Finishing the PieChart, TotalRevenue and the PropertyReferals components.
Using the refine utility for the properties page routing in App.tsx.
Creation of Form component ToDo(need to finish the rest of inputs).
Creation of pages for the properties CRUD.
Creation of the CustomButton using mui.
Changing the bgcolor of the index.tsx layout file line 35.